Remove VariableTransformer super class and move Unifier to //logic/resolvable
