--- logic/transformer/VariableTransformer.java
@@ -1,46 +1,0 @@
-/*
- * Copyright (C) 2020 Grakn Labs
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as
- * published by the Free Software Foundation, either version 3 of the
- * License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <https://www.gnu.org/licenses/>.
- */
-
-package grakn.core.logic.transformer;
-
-import grakn.core.common.exception.GraknException;
-import grakn.core.common.parameters.Label;
-import graql.lang.pattern.variable.Reference;
-
-import java.util.Set;
-
-import static grakn.common.util.Objects.className;
-import static grakn.core.common.exception.ErrorMessage.Pattern.INVALID_CASTING;
-
-public abstract class VariableTransformer {
-
-    public boolean isUnifier() {
-        return false;
-    }
-
-    public Unifier asUnifier() {
-        throw GraknException.of(INVALID_CASTING, className(this.getClass()), className(Unifier.class));
-    }
-
-    public boolean isMapping() {
-        return false;
-    }
-
-    public Mapping asMapped() {
-        throw GraknException.of(INVALID_CASTING, className(this.getClass()), className(Mapping.class));
-    }
-}
